+++ src/mm_2buddy.c
@@ -13,7 +13,7 @@ struct EmptyValue { // needs set: mmInfo; type=t_empty; next; everything else ca
 };
 
 #define  BSZ(X) (1ull<<(X))
-#define BSZI(X) (64-__builtin_clzl((X)-1ull))
+#define BSZI(X) ((u8)(64-__builtin_clzl((X)-1ull)))
 #define  MMI(X) X
 #define   BN(X) b1_##X
 #define buckets b1_buckets
@@ -24,7 +24,7 @@ struct EmptyValue { // needs set: mmInfo; type=t_empty; next; everything else ca
 #undef BSZI
 
 #define  BSZ(X) (3ull<<(X))
-#define BSZI(X) (64-__builtin_clzl((X)/3-1ull))
+#define BSZI(X) ((u8)(64-__builtin_clzl((X)/3-1ull)))
 #define  MMI(X) ((X)|64)
 #define   BN(X) b3_##X
 #define buckets b3_buckets
